Fix Changelog Verification (#19)
This includes some fixes to the changelog tool including verification that no longer accepts everything, requiring the changelog file to have # Change Log at the top, restricting the arbitrary text at the top to be a single line, and adding tests for invalid changelog files. The command will also produce an exit code of 1 if it fails for any reason, unless it is called with zero arguments.