Switch Docker images using musl based changelog-tool
It will be more portable across Linux environments when we want to
grab from an image for use in other containers.

In order to do this, I needed to add a decent amount of the Makefile
changes that are needed to build portable pony binaries.

How tests are run is a bit of a hack and doesn't fully fit in with
the changes, but they work. I'm going to open an issue to clean up
how tests are run.

Closes #35