Refactor handling of unreleased part of changelogs
The new `add` command has resulted in some workarounds for the original
assumption that the "unreleased" release is either nonexistent or empty
in many code paths. This is resolved by adding a separated class,
`Unreleased`, that enforces the requirement that this part of the
changelog must contain all sections if it exists.

Closes #34